backend/payment-service/requirements.txt
flask
mysql-connector-python
pika
requests
flask-cors
